rename api to leaderboard_app
