+++ django_redis_leaderboard/django_redis_leaderboard/settings.py
@@ -87,7 +87,7 @@ STATICFILES_DIRS = (
     # Put strings here, like "/home/html/static" or "C:/www/django/static".
     # Always use forward slashes, even on Windows.
     # Don't forget to use absolute paths, not relative paths.
-    os.path.join(BASE_DIR, "django_leaderboard/static"),    
+    os.path.join(BASE_DIR, "leaderboard_app/static"),    
 )
 
 # List of finder classes that know how to find static files in
@@ -143,7 +143,7 @@ INSTALLED_APPS = (
     'django.contrib.sites',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'django_leaderboard',
+    'leaderboard_app',
     'rest_framework',
 )
