+++ django_redis_leaderboard/django_redis_leaderboard/urls.py
@@ -5,7 +5,7 @@ urlpatterns = [
     # Redirect to leaderboard
     path('', RedirectView.as_view(url='leaderboard/highscores/sample_game/')),
     # Leaderboard
-    path('leaderboard/', include('django_leaderboard.urls')),
+    path('leaderboard/', include('leaderboard_app.urls')),
     # auth support for rest framework
     path('restframework/', include('rest_framework.urls', namespace='rest_framework'))
 ]
