+++ django_redis_leaderboard/django_redis_leaderboard/wsgi.py
@@ -15,7 +15,7 @@ framework.
 """
 import os
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "samplegame.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "django_redis_leaderboard.settings")
 
 # This application object is used by any WSGI server configured to use this
 # file. This includes Django's development server, if the WSGI_APPLICATION
