+++ django_redis_leaderboard/leaderboard_app/static/favicon.ico
