+++ django_redis_leaderboard/leaderboard_app/static/home/css/sidebar.css
