+++ django_redis_leaderboard/leaderboard_app/static/home/javascript/messages.js
