+++ django_redis_leaderboard/leaderboard_app/templates/leaderboard_app/highscores.html
